requirements.txt
absl-py==0.12.0
ansible==4.7.0
ansible-core==2.11.6
antlr4-python3-runtime==4.8
appdirs==1.4.4
arrow==1.2.0
asgiref==3.3.1
astroid==2.4.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
binaryornot==0.4.4
black==21.12b0
bs4==0.0.1
cachetools==4.2.1
Cerberus==1.3.4
certifi==2020.12.5
cffi==1.14.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.3
click-completion==0.5.2
click-help-colors==0.9.1
cloudpickle==2.0.0
colorama==0.4.4
configparser==5.1.0
cookiecutter==1.7.3
crayons==0.4.0
cryptography==36.0.0
cycler==0.10.0
Cython==0.29.26
dataclasses==0.6
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.4.2
Django==3.1.7
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-filter==2.4.0
django-taggit==2.0.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
docker==5.0.3
easyocr==1.4.1
et-xmlfile==1.1.0
fake-useragent==0.1.11
fasteners==0.16.3
filelock==3.0.12
flatbuffers==1.12
fonttools==4.28.4
future==0.18.2
fvcore==0.1.5.post20211023
gast==0.3.3
gnureadline==8.0.0
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h11==0.12.0
h5py==2.10.0
html5lib==1.1
hydra-core==1.1.1
idna==2.10
imageio==2.9.0
imantics==0.1.12
imgaug==0.4.0
importlib-metadata==1.5.0
importlib-resources==5.4.0
imutils==0.5.4
instagram-private-api==1.6.0.0
install==1.3.4
iopath==0.1.9
isort==4.3.21
Jinja2==3.0.2
jinja2-time==0.2.0
jsonschema==3.2.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
labelme2coco==0.1.2
lazy-object-proxy==1.4.3
libclang==12.0.0
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.1
mccabe==0.6.1
molecule==3.0.6
more-itertools==8.2.0
mypy-extensions==0.4.3
networkx==2.5.1
numpy==1.19.5
oauthlib==3.1.0
omegaconf==2.1.1
opencv-python==4.5.1.48
opencv-python-headless==4.5.4.60
openpyxl==3.0.9
opt-einsum==3.3.0
outcome==1.1.0
packaging==20.1
pandas==1.3.4
paramiko==2.8.0
pathspec==0.9.0
pexpect==4.8.0
Pillow==8.2.0
pixellib==0.6.1
platformdirs==2.4.0
pluggy==0.13.1
portalocker==2.3.2
poyo==0.5.0
prettyprint==0.1.5
protobuf==3.15.8
psycopg2-binary==2.9.1
ptyprocess==0.7.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.2
PyJWT==1.5.2
pylint==2.5.2
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.17.3
PySocks==1.7.1
pytesseract==0.3.8
pytest==5.3.5
pytest-django==3.8.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-dotenv==0.13.0
python-gilt==1.2.3
python-slugify==5.0.2
python-telegram-bot==12.7
python3-openid==3.2.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
resolvelib==0.5.4
rsa==4.7.2
scikit-image==0.18.1
scipy==1.6.2
selenium==4.1.0
sh==1.13.1
Shapely==1.7.1
shellingham==1.4.0
six==1.15.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sorl-thumbnail==12.6.3
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.1
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
text-unidecode==1.3
thread6==0.2.0
tifffile==2021.4.8
toml==0.10.1
tomli==1.2.3
torch==1.10.0
torchvision==0.11.1
tornado==6.0.4
tqdm==4.62.3
tree-format==0.1.2
trio==0.19.0
trio-websocket==0.9.2
twilio==6.45.1
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.0.26
wcwidth==0.1.8
webdriver-manager==3.5.2
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==1.0.1
wrapt==1.12.1
wsproto==1.0.0
xlrd==2.0.1
XlsxWriter==3.0.2
xmljson==0.2.1
yacs==0.1.8
yamllint==1.26.3
zipp==3.6.0
